Add append option to unsupported features
This could be done, but ideally an option should be added to
asyncrun.vim, now it would require setting the system-wide option
`g:asynrun_trim`. It would also require doing a PR request so we can
drop the hacks to remove the extra lines asyncrun adds to the quickfix
list. The current hacks used to circumvent that issue would be
problematic when using the append feature.